+++ projects/exponent/index.js
@@ -1,18 +1,13 @@
-const { getConnection, getProvider } = require('../helper/solana')
-const { PublicKey } = require('@solana/web3.js')
-const { decodeAccount } = require('../helper/utils/solana/layout')
-const { Program } = require('@coral-xyz/anchor')
+const { getConnection, getProvider } = require("../helper/solana");
+const { PublicKey } = require("@solana/web3.js");
+const { decodeAccount } = require("../helper/utils/solana/layout");
+const { Program } = require("@coral-xyz/anchor");
 
 const { getConfig } = require('../helper/cache')
 const idl = require('./idl.json')
-const vaultsIdl = require('./vaults_idl.json')
 
-const SYNTHETIC_MINT_MAP = {
-  'USD1111111111111111111111111111111111111111': 'EPjFWdd5AufqSSqeM2qN1xzybapC8G4wEGGkZwyTDt1v', // synthetic USD -> USDC
-  'USD1111111111111111111111111111111111111119': 'DEkqHyPN7GMRJ5cArtQFAWefqbZb33Hyf6s5iCwjEonT', // synthetic USD9 -> USDe
-}
 
-async function addCoreTvl(api) {
+async function tvl(api) {
   const provider = getProvider()
   const connection = getConnection()
   
@@ -59,26 +54,8 @@ async function addCoreTvl(api) {
   }
 }
 
-async function addManagedVaultAum(api) {
-  const program = new Program(vaultsIdl, getProvider())
-  const vaults = await program.account.exponentStrategyVault.all()
-  for (const { account } of vaults) {
-    // skip non-active vaults that aren't shown on the UI
-    if (account.statusFlags === 8) continue
-    const total = BigInt(account.financials.aumInBase.toString()) + BigInt(account.financials.aumInBaseInPositions.toString())
-    if (total <= 0n) continue
-    const mint = account.underlyingMint.toBase58()
-    api.add(SYNTHETIC_MINT_MAP[mint] || mint, total)
-  }
-}
-
-async function tvl(api) {
-  await addCoreTvl(api)
-  await addManagedVaultAum(api)
-}
-
 module.exports = {
   timetravel: false,
-  methodology: 'Core TVL: sum the total supply of each Exponent wrapped yield-bearing token multiplied by its on-chain last-seen SY exchange rate, emitted under the underlying mint from the Exponent API. Managed vault TVL: read each strategy vault\'s on-chain `aumInBase` (idle reserves) and `aumInBaseInPositions` (value of deployed positions).',
+  methodology: "TVL is calculated by summing the total supply of each Exponent wrapped Yield bearing token and multiplying their base asset amount by the price of the underlying token",
   solana: { tvl },
-}
+};
